refactor(alignments): absolute genomic uint32 coords end-to-end
Every position array emitted by the alignments worker (reads, gaps,
mismatches, interbase, softclip bases, modifications, SNP/noncov/indicator/
modCov segments, sashimi junctions, arcs, connecting lines, coverage
start offset) now stores absolute genomic uint32. No `regionStart +`
arithmetic remains in any consumer — SVG export, Canvas2D, hit testing,
tooltips, main-thread layout, and model views all consume absolute coords
directly.

Rationale: GPU smooth zoom invalidates any anchor-relative convention when
the region boundary shifts; genomic positions are always >=0 so no uint32
wrapping trick is needed; reversed regions are handled by the drawing
layer (flipX on GPU, bpToScreenX on Canvas2D) independent of the coord
convention; float32 cannot hold 3 Gbp positions (~256 bp precision loss)
so uint32 is the correct storage type.

All GPU shaders use hp-math: each consumes uint absolute positions and
converts to clip space via hpClipX(hpSplitUint(absPos)) or hpLinear for
curve math. The arc shader uses hpLinear-derived normalized [0,1]
fractions for bezier interpolation to preserve precision through the
curve. regionStart, domainStart, and domainEnd are removed from the
alignments UBO.

sashimiX1/X2 and arcX1/X2 converted Float32Array -> Uint32Array to avoid
the 3 Gbp float32 precision trap. computeInsertionIndicators no longer
takes regionStart (implicit via coverageStartOffset).

Synteny coverage renderer + packers updated to drop their regionStart
add-backs since computeSNPCoverage/computeInsertionIndicators now return
absolute positions and coverage.startOffset is absolute.

Added plugins/alignments/src/shared/absoluteCoords.test.ts verifying
forward/reversed pixel and clip-space mapping agree and that Canvas2D
and GPU pipelines produce consistent visual coordinates.

Documented the convention in agent-docs/ARCHITECTURE.md ("Coordinate
convention"), the project CLAUDE.md, and the PileupDataResult type
docstring. Existing unit tests (596) pass; jbrowse-web image-snapshot
tests diverge and are left for regeneration.

+++ plugins/alignments/src/LinearAlignmentsDisplay/components/rendererTypes.ts
@@ -134,8 +134,8 @@ export interface ModCoverageUploadData {
 
 export interface ArcsUploadData {
   regionStart: number
-  arcX1: Float32Array
-  arcX2: Float32Array
+  arcX1: Uint32Array
+  arcX2: Uint32Array
   arcColorTypes: Float32Array
   arcIsArc: Uint8Array
   numArcs: number
